Move TenancyFilterSet to filters.py
